Fixed the byte to be unsigned char instead of unsigned short. The initial idea was for both of them to be 16bit, so that we can eliminate type casting, but I started implementing the Arithmethic Instructions and I need the bytes to overflow, so I can set the Flags in the F register correctly

+++ NaughtyGameboy/Source/BitUtil.h
@@ -29,12 +29,12 @@ inline byte GetHighByte(ushort src)
 
 inline void SetLowByte(ushort* dst, byte value)
 {
-	byte high = GetHighByte(*dst);
+	ushort high = GetHighByte(*dst);
 	*dst = (high << 8) | value;
 }
 
 inline void SetHighByte(ushort* dst, byte value)
 {
-	byte low = GetLowByte(*dst);
+	ushort low = GetLowByte(*dst);
 	*dst = (value << 8) | low;
 }
